Setup cinder block devices to use with ceph internally for testing
I had issues getting loopback devices to work with ceph-volume, so this is a
nasty yet somewhat elegant workaround for it. This uses a ruby script on the VM
that utilizes fog-openstack to create five block devices and attaches them to
the VM. If it finds any current available volumes matching our naming scheme,
remove them since we can't clear them after a kitchen destroy.