Implement ceph_keyring resource
This creates a ceph_keyring resource which simply adds a keyring file with a
given key. This is a more simple approach than what ceph_chef_client provides
which requires a connection to a ceph server. With this use case, we'll be
creating the auth tokens manually and putting them (likely) in encrypted
data bags and using this resource to add it.

This is needed in situations where you only need to provide the auth token and
nothing else.