Major refactor into a resource driven cookbook
This removes our dependency on the legacy ceph-chef cookbook that we also
forked and migrates us to having a proper resource driven cookbook to make it
easier to use in general.

I tried my best to not change how this works for clients simply needing to
include the default recipe. However, there will need to be changes in a few
places to make this work properly.

I also refactored and updated the multi-node setup to make it work with newer
Terraform. In addition, I created a new ceph_test resource to ease with setting
up a single local instance of a Ceph cluster for test-kitchen use.

This is the first step in getting us to newer versions of Ceph along with adding
AlmaLinux 8 and beyond.

Signed-off-by: Lance Albertson <lance@osuosl.org>